StatisticsPage: fix:연관관계 문제 해결 fetch join

+++ src/main/java/com/sollertia/habit/domain/habit/dto/HabitSummaryResponseVo.java
@@ -1,7 +1,7 @@
 package com.sollertia.habit.domain.habit.dto;
 
+import com.sollertia.habit.domain.category.enums.Category;
 import com.sollertia.habit.domain.habit.Habit;
-import com.sollertia.habit.domain.habit.enums.Category;
 import lombok.experimental.SuperBuilder;
 
 @SuperBuilder
@@ -10,7 +10,7 @@ public class HabitSummaryResponseVo {
     private String title;
     private String durationStart;
     private String durationEnd;
-    private Long current;
+    private int current;
     private Long sessionDuration;
     private Long progressPercentage;
     private Category category;
